Add files via upload
This C program takes a cash amount as input and breaks it down into the smallest number of bills and coins based on Philippine currency denominations (₱1000, ₱500, ₱100, ₱50, ₱20, ₱10, ₱5, and ₱1). It uses integer division and modulus operations to compute how many of each denominations are needed, then it displays the result in a structured format.